Backend-MS3/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose==3.3.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
email-validator==2.1.0.post1
requests==2.31.0
cryptography==41.0.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

